♻️ : 댓글 관련 재구조화
